refactor: di에 맞춰 생성자 정리 및 폴더구조 MVVM 정리
