+++ lib/presentation/component/recipe_thumbnail_card.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:recipe_app/presentation/ui/color_styles.dart';
 import 'package:recipe_app/presentation/ui/text_styles.dart';
 
-import '../../data/model/recipe.dart';
+import '../../domain/model/recipe.dart';
 
 class RecipeThumbnailCard extends StatelessWidget {
   final Recipe recipe;
