+++ lib/presentation/saved_recipes/saved_recipes_state.dart
@@ -1,5 +1,5 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
-import '../../data/model/recipe.dart';
+import '../../domain/model/recipe.dart';
 
 part 'saved_recipes_state.freezed.dart';
