+++ lib/presentation/saved_recipes/saved_recipes_view_model.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:recipe_app/presentation/saved_recipes/saved_recipes_state.dart';
 
-import '../../data/model/recipe.dart';
-import '../../data/repository/recipe_ropository.dart';
+import '../../domain/model/recipe.dart';
+import '../../domain/repository/recipe_ropository.dart';
 
 class SavedRecipesViewModel with ChangeNotifier {
   final RecipeRepository _recipeRepository;
@@ -12,7 +12,8 @@ class SavedRecipesViewModel with ChangeNotifier {
 
   SavedRecipesState get state => _state;
 
-  SavedRecipesViewModel(this._recipeRepository);
+  SavedRecipesViewModel({required RecipeRepository recipeRepository})
+  : _recipeRepository = recipeRepository;
 
   void fetchRecipes() async {
     _state = state.copyWith(isLoading: true);
