+++ lib/presentation/search_recipes/search_recipes_state.dart
@@ -1,6 +1,6 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
 
-import '../../data/model/recipe.dart';
+import '../../domain/model/recipe.dart';
 
 part 'search_recipes_state.freezed.dart';
