+++ lib/presentation/search_recipes/search_recipes_view_model.dart
@@ -1,24 +1,25 @@
 import 'package:flutter/material.dart';
-import 'package:recipe_app/data/repository/recipe_ropository.dart';
+import 'package:recipe_app/domain/repository/recipe_ropository.dart';
 import 'package:recipe_app/presentation/search_recipes/search_recipes_state.dart';
 
-import '../../data/model/recipe.dart';
+import '../../domain/model/recipe.dart';
 
 class SearchRecipesViewModel with ChangeNotifier {
-  final RecipeRepository recipeRepository;
+  final RecipeRepository _recipeRepository;
 
   //상태
   SearchRecipesState _state = const SearchRecipesState();
 
-  SearchRecipesState get state => _state;
+  SearchRecipesViewModel({required RecipeRepository recipeRepository})
+    : _recipeRepository = recipeRepository;
 
-  SearchRecipesViewModel({required this.recipeRepository});
+  SearchRecipesState get state => _state;
 
   void fetchRecipes() async {
     _state = state.copyWith(isLoading: true);
     notifyListeners();
 
-    List<Recipe> recipes = await recipeRepository.getRecipes();
+    List<Recipe> recipes = await _recipeRepository.getRecipes();
     _state = state.copyWith(
       recipes: recipes,
       filteredRecipes: recipes,
@@ -28,13 +29,13 @@ class SearchRecipesViewModel with ChangeNotifier {
   }
 
   void search(String keyword) async {
-
     _state = state.copyWith(isLoading: true);
     notifyListeners();
 
-    final results = state.recipes.where((recipe) {
-      return recipe.recipeName.contains(keyword);
-    }).toList();
+    final results =
+        state.recipes.where((recipe) {
+          return recipe.recipeName.contains(keyword);
+        }).toList();
 
     _state = state.copyWith(
       searchKeyword: keyword,
@@ -43,5 +44,4 @@ class SearchRecipesViewModel with ChangeNotifier {
     );
     notifyListeners();
   }
-
 }
